requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langchain-google-genai
langchain-community
mangum
python-dotenv
